+++ CHANGELOG.md
@@ -7,7 +7,9 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Add Serde `Deserialize` and `Serialize` support for `CpuQuantity` and `MemoryQuantity` ([#724]).
+- Add `DynamicValues` struct to work with operator `values.yaml` files during runtime ([#723]).
 
+[#723]: https://github.com/stackabletech/operator-rs/pull/723
 [#724]: https://github.com/stackabletech/operator-rs/pull/724
 
 ## [0.62.0] - 2024-01-19